requirements.txt
fastapi
uvicorn
jinja2
pymongo
authlib
itsdangerous
httpx
python-dotenv
python-multipart
cloudinary